refactor(webui): extract event projection helpers
